Updated code to extract the identities logic from the secureenclave file
